refactor: extract PaginationCoordinator from MainContentCoordinator
