+++ TablePro/Views/Main/MainContentCoordinator.swift
@@ -101,6 +101,7 @@ final class MainContentCoordinator {
 
     @ObservationIgnored private(set) var filterCoordinator: FilterCoordinator!
     @ObservationIgnored private(set) var queryExecutionCoordinator: QueryExecutionCoordinator!
+    @ObservationIgnored private(set) var paginationCoordinator: PaginationCoordinator!
 
     /// Stable identifier for this coordinator's window (set by MainContentView on appear)
     var windowId: UUID?
@@ -397,6 +398,7 @@ final class MainContentCoordinator {
 
         self.filterCoordinator = FilterCoordinator(parent: self)
         self.queryExecutionCoordinator = QueryExecutionCoordinator(parent: self)
+        self.paginationCoordinator = PaginationCoordinator(parent: self)
 
         Self.lifecycleLogger.info(
             "[open] MainContentCoordinator.init done connId=\(connection.id, privacy: .public) elapsedMs=\(Int(Date().timeIntervalSince(initStart) * 1_000))"
